feat: implement http handlers, api service, preseneters and swagger docs

+++ internal/user/port/user.go
@@ -8,6 +8,6 @@ import (
 type Repo interface {
 	Create(ctx context.Context, user domain.User) (domain.APIKey, error)
 	GetByID(ctx context.Context, ID uint) (domain.User, error)
-	CreditBalance(ctx context.Context, ID uint, amount int64) (domain.User, error)
-	DebitBalance(ctx context.Context, ID uint, amount int64) (domain.User, error)
+	CreditBalance(ctx context.Context, ID uint, amount float64) (domain.User, error)
+	DebitBalance(ctx context.Context, ID uint, amount float64) (domain.User, error)
 }
